+++ src/screens/client/proposals/ProposalInfo.js
