+++ src/screens/client/serviceDesk/NewTicket.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
-import { useAuth } from "../../context/AuthContext";
-import { BASE_URL } from "../../utils/endPointNames.js";
+import { useAuth } from "../../../context/AuthContext";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 import toast from "react-hot-toast";
 
 const CreateTicket = () => {
