+++ src/screens/client/serviceDesk/ViewTicket.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import axios from "axios";
-import { useAuth } from "../../context/AuthContext";
-import { BASE_URL } from "../../utils/endPointNames";
+import { useAuth } from "../../../context/AuthContext.jsx";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 
 const ViewTicket = () => {
   const [auth] = useAuth();
